pyproject.toml
[project]
name = "payoutsystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "drf-spectacular-sidecar>=2025.12.1",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.1",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
]
